Beginning world rendering and terrain generation
- Started work on the terrain generation; this uses fastnoise2's 3D
  noise generation features

- Basic chunk renderer. This supports a moving range of chunks rendered
  around the central chunk at the current position. It definitely needs
  work though lmao

+++ src/world/chunk/Chunk.h
@@ -76,7 +76,7 @@ struct Chunk {
     /**
      * X/Z coordinates of this chunk, in world chunk coordinate space.
      */
-    glm::vec2 worldPos;
+    glm::ivec2 worldPos;
 
     /**
      * Chunk slice pointers for each horizontal layer of the chunk. If there are no blocks at that
